Clean up client-side saved object data loading infrastructure (#30241)
