+++ src/fixtures/stubbed_saved_object_index_pattern.js
@@ -18,13 +18,13 @@
  */
 
 import FixturesLogstashFieldsProvider from 'fixtures/logstash_fields';
-import { SavedObject } from 'ui/saved_objects';
+import { SimpleSavedObject } from 'ui/saved_objects';
 
 export function FixturesStubbedSavedObjectIndexPatternProvider(Private) {
   const mockLogstashFields = Private(FixturesLogstashFieldsProvider);
 
   return function (id) {
-    return new SavedObject(undefined, {
+    return new SimpleSavedObject(undefined, {
       id,
       type: 'index-pattern',
       attributes: {
