+++ src/legacy/core_plugins/kibana/public/dashboard/saved_dashboard/saved_dashboard.js
@@ -21,7 +21,7 @@ import angular from 'angular';
 import { uiModules } from 'ui/modules';
 import { createDashboardEditUrl } from '../dashboard_constants';
 import { createLegacyClass } from 'ui/utils/legacy_class';
-import { SavedObjectProvider } from 'ui/courier';
+import { SavedObjectProvider } from 'ui/saved_objects/saved_object';
 import {
   extractReferences,
   injectReferences,
