+++ src/legacy/core_plugins/kibana/public/dashboard/saved_dashboard/saved_dashboards.js
@@ -20,9 +20,8 @@
 import { i18n } from '@kbn/i18n';
 import './saved_dashboard';
 import { uiModules } from 'ui/modules';
-import { SavedObjectLoader } from 'ui/courier/saved_object/saved_object_loader';
+import { SavedObjectLoader, SavedObjectsClientProvider } from 'ui/saved_objects';
 import { savedObjectManagementRegistry } from '../../management/saved_object_registry';
-import { SavedObjectsClientProvider } from 'ui/saved_objects';
 
 const module = uiModules.get('app/dashboard');
 
@@ -38,5 +37,5 @@ savedObjectManagementRegistry.register({
 // This is the only thing that gets injected into controllers
 module.service('savedDashboards', function (Private, SavedDashboard, kbnIndex, kbnUrl, $http, chrome) {
   const savedObjectClient = Private(SavedObjectsClientProvider);
-  return new SavedObjectLoader(SavedDashboard, kbnIndex, kbnUrl, $http, chrome, savedObjectClient);
+  return new SavedObjectLoader(SavedDashboard, kbnUrl, chrome, savedObjectClient);
 });
