+++ src/legacy/core_plugins/kibana/public/discover/saved_searches/_saved_search.js
@@ -20,7 +20,7 @@
 import 'ui/notify';
 import { uiModules } from 'ui/modules';
 import { createLegacyClass } from 'ui/utils/legacy_class';
-import { SavedObjectProvider } from 'ui/courier';
+import { SavedObjectProvider } from 'ui/saved_objects/saved_object';
 
 const module = uiModules.get('discover/saved_searches', [
   'kibana/notify',
