+++ src/legacy/core_plugins/kibana/public/discover/saved_searches/saved_searches.js
@@ -20,9 +20,8 @@
 import './_saved_search';
 import 'ui/notify';
 import { uiModules } from 'ui/modules';
-import { SavedObjectLoader } from 'ui/courier/saved_object/saved_object_loader';
+import { SavedObjectLoader, SavedObjectsClientProvider } from 'ui/saved_objects';
 import { savedObjectManagementRegistry } from '../../management/saved_object_registry';
-import { SavedObjectsClientProvider } from 'ui/saved_objects';
 const module = uiModules.get('discover/saved_searches', [
   'kibana/notify'
 ]);
@@ -36,7 +35,7 @@ savedObjectManagementRegistry.register({
 
 module.service('savedSearches', function (Private, Promise, config, kbnIndex, createNotifier, SavedSearch, kbnUrl, $http, chrome) {
   const savedObjectClient = Private(SavedObjectsClientProvider);
-  const savedSearchLoader = new SavedObjectLoader(SavedSearch, kbnIndex, kbnUrl, $http, chrome, savedObjectClient);
+  const savedSearchLoader = new SavedObjectLoader(SavedSearch, kbnUrl, chrome, savedObjectClient);
   // Customize loader properties since adding an 's' on type doesn't work for type 'search' .
   savedSearchLoader.loaderProperties = {
     name: 'searches',
