+++ src/legacy/core_plugins/kibana/public/visualize/saved_visualizations/_saved_vis.js
@@ -30,7 +30,7 @@ import { uiModules } from 'ui/modules';
 import { updateOldState } from 'ui/vis/vis_update_state';
 import { VisualizeConstants } from '../visualize_constants';
 import { createLegacyClass } from 'ui/utils/legacy_class';
-import { SavedObjectProvider } from 'ui/courier';
+import { SavedObjectProvider } from 'ui/saved_objects/saved_object';
 import {
   extractReferences,
   injectReferences,
