+++ src/legacy/ui/public/saved_objects/saved_object.js
@@ -31,20 +31,21 @@
 import angular from 'angular';
 import _ from 'lodash';
 
-import { InvalidJSONProperty, SavedObjectNotFound } from '../../errors';
-import MappingSetupProvider from '../../utils/mapping_setup';
+import { InvalidJSONProperty, SavedObjectNotFound } from '../errors';
+import MappingSetupProvider from '../utils/mapping_setup';
 
-import { SearchSourceProvider } from '../search_source';
-import { SavedObjectsClientProvider, findObjectByTitle } from '../../saved_objects';
-import { migrateLegacyQuery } from '../../utils/migrate_legacy_query';
-import { recentlyAccessed } from '../../persisted_log';
+import { SearchSourceProvider } from '../courier/search_source';
+import { findObjectByTitle } from './find_object_by_title';
+import { SavedObjectsClientProvider } from './saved_objects_client_provider';
+import { migrateLegacyQuery } from '../utils/migrate_legacy_query';
+import { recentlyAccessed } from '../persisted_log';
 import { i18n } from '@kbn/i18n';
 
 /**
  * An error message to be used when the user rejects a confirm overwrite.
  * @type {string}
  */
-const OVERWRITE_REJECTED = i18n.translate('common.ui.courier.savedObject.overwriteRejectedDescription', {
+const OVERWRITE_REJECTED = i18n.translate('common.ui.savedObjects.overwriteRejectedDescription', {
   defaultMessage: 'Overwrite confirmation was rejected'
 });
 
@@ -52,7 +53,7 @@ const OVERWRITE_REJECTED = i18n.translate('common.ui.courier.savedObject.overwri
  * An error message to be used when the user rejects a confirm save with duplicate title.
  * @type {string}
  */
-const SAVE_DUPLICATE_REJECTED = i18n.translate('common.ui.courier.savedObject.saveDuplicateRejectedDescription', {
+const SAVE_DUPLICATE_REJECTED = i18n.translate('common.ui.savedObjects.saveDuplicateRejectedDescription', {
   defaultMessage: 'Save with duplicate title confirmation was rejected'
 });
 
@@ -70,6 +71,15 @@ export function SavedObjectProvider(Promise, Private, Notifier, confirmModalProm
   const SearchSource = Private(SearchSourceProvider);
   const mappingSetup = Private(MappingSetupProvider);
 
+  /**
+   * The SavedObject class is a base class for saved objects loaded from the server and
+   * provides additional functionality besides loading/saving/deleting/etc.
+   *
+   * It is overloaded and configured to provide type-aware functionality.
+   * To just retrieve the attributes of saved objects, it is recommended to use SavedObjectLoader
+   * which returns instances of SimpleSavedObject which don't introduce additional type-specific complexity.
+   * @param {*} config
+   */
   function SavedObject(config) {
     if (!_.isObject(config)) config = {};
 
@@ -388,16 +398,16 @@ export function SavedObjectProvider(Promise, Private, Notifier, confirmModalProm
         .catch(err => {
           // record exists, confirm overwriting
           if (_.get(err, 'res.status') === 409) {
-            const confirmMessage = i18n.translate('common.ui.courier.savedObject.confirmModal.overwriteConfirmationMessage', {
+            const confirmMessage = i18n.translate('common.ui.savedObjects.confirmModal.overwriteConfirmationMessage', {
               defaultMessage: 'Are you sure you want to overwrite {title}?',
               values: { title: this.title }
             });
 
             return confirmModalPromise(confirmMessage, {
-              confirmButtonText: i18n.translate('common.ui.courier.savedObject.confirmModal.overwriteButtonLabel', {
+              confirmButtonText: i18n.translate('common.ui.savedObjects.confirmModal.overwriteButtonLabel', {
                 defaultMessage: 'Overwrite',
               }),
-              title: i18n.translate('common.ui.courier.savedObject.confirmModal.overwriteTitle', {
+              title: i18n.translate('common.ui.savedObjects.confirmModal.overwriteTitle', {
                 defaultMessage: 'Overwrite {name}?',
                 values: { name: this.getDisplayName() }
               }),
@@ -410,13 +420,13 @@ export function SavedObjectProvider(Promise, Private, Notifier, confirmModalProm
     };
 
     const displayDuplicateTitleConfirmModal = () => {
-      const confirmMessage = i18n.translate('common.ui.courier.savedObject.confirmModal.saveDuplicateConfirmationMessage', {
+      const confirmMessage = i18n.translate('common.ui.savedObjects.confirmModal.saveDuplicateConfirmationMessage', {
         defaultMessage: `A {name} with the title '{title}' already exists. Would you like to save anyway?`,
         values: { title: this.title, name: this.getDisplayName() }
       });
 
       return confirmModalPromise(confirmMessage, {
-        confirmButtonText: i18n.translate('common.ui.courier.savedObject.confirmModal.saveDuplicateButtonLabel', {
+        confirmButtonText: i18n.translate('common.ui.savedObjects.confirmModal.saveDuplicateButtonLabel', {
           defaultMessage: 'Save {name}',
           values: { name: this.getDisplayName() }
         })
