+++ src/legacy/ui/public/saved_objects/saved_objects_client.ts
@@ -31,7 +31,7 @@ import { CreateResponse, FindOptions, UpdateResponse } from '../../../server/sav
 import { isAutoCreateIndexError, showAutoCreateIndexErrorPage } from '../error_auto_create_index';
 import { kfetch, KFetchQuery } from '../kfetch';
 import { keysToCamelCaseShallow, keysToSnakeCaseShallow } from '../utils/case_conversion';
-import { SavedObject } from './saved_object';
+import { SimpleSavedObject } from './simple_saved_object';
 
 interface RequestParams {
   method: 'POST' | 'GET' | 'PUT' | 'DELETE';
@@ -60,7 +60,7 @@ interface UpdateOptions {
 }
 
 interface BatchResponse<T extends SavedObjectAttributes = SavedObjectAttributes> {
-  savedObjects: Array<SavedObject<T>>;
+  savedObjects: Array<SimpleSavedObject<T>>;
 }
 
 interface FindResults<T extends SavedObjectAttributes = SavedObjectAttributes>
@@ -73,7 +73,9 @@ interface FindResults<T extends SavedObjectAttributes = SavedObjectAttributes>
 interface BatchQueueEntry {
   type: string;
   id: string;
-  resolve: <T extends SavedObjectAttributes>(value: SavedObject<T> | PlainSavedObject<T>) => void;
+  resolve: <T extends SavedObjectAttributes>(
+    value: SimpleSavedObject<T> | PlainSavedObject<T>
+  ) => void;
   reject: (reason?: any) => void;
 }
 
@@ -91,6 +93,14 @@ const BATCH_INTERVAL = 100;
 
 const API_BASE_URL = '/api/saved_objects/';
 
+/**
+ * The SavedObjectsClient class acts as a generic data fetcher
+ * and data saver for saved objects regardless of type.
+ *
+ * If possible, this class should be used to load saved objects
+ * instead of the SavedObjectLoader class which implements some
+ * additional functionality.
+ */
 export class SavedObjectsClient {
   /**
    * Throttled processing of get requests into bulk requests at 100ms interval
@@ -145,7 +155,7 @@ export class SavedObjectsClient {
     type: string,
     attributes: T,
     options: CreateOptions = {}
-  ): Promise<SavedObject<T>> => {
+  ): Promise<SimpleSavedObject<T>> => {
     if (!type || !attributes) {
       return Promise.reject(new Error('requires type and attributes'));
     }
@@ -257,7 +267,7 @@ export class SavedObjectsClient {
   public get = <T extends SavedObjectAttributes>(
     type: string,
     id: string
-  ): Promise<SavedObject<T>> => {
+  ): Promise<SimpleSavedObject<T>> => {
     if (!type || !id) {
       return Promise.reject(new Error('requires type and id'));
     }
@@ -311,7 +321,7 @@ export class SavedObjectsClient {
     id: string,
     attributes: T,
     { version, migrationVersion, references }: UpdateOptions = {}
-  ): Promise<SavedObject<T>> {
+  ): Promise<SimpleSavedObject<T>> {
     if (!type || !id || !attributes) {
       return Promise.reject(new Error('requires type, id and attributes'));
     }
@@ -336,8 +346,8 @@ export class SavedObjectsClient {
 
   private createSavedObject<T extends SavedObjectAttributes>(
     options: PlainSavedObject<T>
-  ): SavedObject<T> {
-    return new SavedObject(this, options);
+  ): SimpleSavedObject<T> {
+    return new SimpleSavedObject(this, options);
   }
 
   private getPath(path: Array<string | undefined>): string {
