+++ src/legacy/ui/public/saved_objects/ui/saved_object_save_as_checkbox.js
@@ -17,4 +17,18 @@
  * under the License.
  */
 
-export { SavedObjectProvider } from './saved_object';
+import { uiModules } from '../../modules';
+import saveObjectSaveAsCheckboxTemplate from './saved_object_save_as_checkbox.html';
+
+uiModules
+  .get('kibana')
+  .directive('savedObjectSaveAsCheckBox', function () {
+    return {
+      restrict: 'E',
+      template: saveObjectSaveAsCheckboxTemplate,
+      replace: true,
+      scope: {
+        savedObject: '='
+      }
+    };
+  });
