+++ x-pack/plugins/maps/public/angular/services/gis_map_saved_object_loader.js
@@ -6,9 +6,8 @@
 
 import './saved_gis_map';
 import { uiModules } from 'ui/modules';
-import { SavedObjectLoader } from 'ui/courier/saved_object/saved_object_loader';
+import { SavedObjectLoader, SavedObjectsClientProvider } from 'ui/saved_objects';
 import { SavedObjectRegistryProvider } from 'ui/saved_objects/saved_object_registry';
-import { SavedObjectsClientProvider } from 'ui/saved_objects';
 
 const module = uiModules.get('app/maps');
 
@@ -22,5 +21,5 @@ SavedObjectRegistryProvider.register({
 // This is the only thing that gets injected into controllers
 module.service('gisMapSavedObjectLoader', function (Private, SavedGisMap, kbnIndex, kbnUrl, $http, chrome) {
   const savedObjectClient = Private(SavedObjectsClientProvider);
-  return new SavedObjectLoader(SavedGisMap, kbnIndex, kbnUrl, $http, chrome, savedObjectClient);
+  return new SavedObjectLoader(SavedGisMap, kbnUrl, chrome, savedObjectClient);
 });
