+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -299,15 +299,6 @@
     "common.ui.courier.requestTimeDescription": "请求从浏览器到 Elasticsearch 以及返回的时间。不包括请求在队列中等候的时间。",
     "common.ui.courier.requestTimeLabel": "请求时间",
     "common.ui.courier.requestTimeValue": "{requestTime}ms",
-    "common.ui.courier.savedObject.confirmModal.overwriteButtonLabel": "覆盖",
-    "common.ui.courier.savedObject.confirmModal.overwriteConfirmationMessage": "确定要覆盖 “{title}”？",
-    "common.ui.courier.savedObject.confirmModal.overwriteTitle": "覆盖“{name}”？",
-    "common.ui.courier.savedObject.confirmModal.saveDuplicateButtonLabel": "保存“{name}”",
-    "common.ui.courier.savedObject.confirmModal.saveDuplicateConfirmationMessage": "具有标题 “{title}” 的 “{name}” 已存在。是否确定要保存？",
-    "common.ui.courier.savedObject.howToSaveAsNewDescription": "在 Kibana 的以前版本中，更改 {savedObjectName} 的名称将创建具有新名称的副本。使用“另存为新的 {savedObjectName}” 复选框可立即达到此目的。",
-    "common.ui.courier.savedObject.overwriteRejectedDescription": "已拒绝覆盖确认",
-    "common.ui.courier.savedObject.saveAsNewLabel": "另存为新的 {savedObjectName}",
-    "common.ui.courier.savedObject.saveDuplicateRejectedDescription": "已拒绝使用重复标题保存确认",
     "common.ui.directives.confirmClickButtonLabel": "是否确定？",
     "common.ui.directives.fieldNameIcons.booleanAriaLabel": "布尔字段",
     "common.ui.directives.fieldNameIcons.conflictFieldAriaLabel": "冲突字段",
@@ -586,8 +577,17 @@
     "common.ui.savedObjectFinder.sortByButtonLabeDescendingScreenReaderOnly": "降序",
     "common.ui.savedObjectFinder.sortByButtonLabel": "名称",
     "common.ui.savedObjectFinder.sortByButtonLabelScreenReaderOnly": "排序依据",
+    "common.ui.savedObjects.confirmModal.overwriteButtonLabel": "覆盖",
+    "common.ui.savedObjects.confirmModal.overwriteConfirmationMessage": "确定要覆盖 “{title}”？",
+    "common.ui.savedObjects.confirmModal.overwriteTitle": "覆盖“{name}”？",
+    "common.ui.savedObjects.confirmModal.saveDuplicateButtonLabel": "保存“{name}”",
+    "common.ui.savedObjects.confirmModal.saveDuplicateConfirmationMessage": "具有标题 “{title}” 的 “{name}” 已存在。是否确定要保存？",
     "common.ui.savedObjects.finder.searchPlaceholder": "搜索……",
     "common.ui.savedObjects.finder.titleLabel": "标题",
+    "common.ui.savedObjects.howToSaveAsNewDescription": "在 Kibana 的以前版本中，更改 {savedObjectName} 的名称将创建具有新名称的副本。使用“另存为新的 {savedObjectName}” 复选框可立即达到此目的。",
+    "common.ui.savedObjects.overwriteRejectedDescription": "已拒绝覆盖确认",
+    "common.ui.savedObjects.saveAsNewLabel": "另存为新的 {savedObjectName}",
+    "common.ui.savedObjects.saveDuplicateRejectedDescription": "已拒绝使用重复标题保存确认",
     "common.ui.savedObjects.saveModal.cancelButtonLabel": "取消",
     "common.ui.savedObjects.saveModal.confirmSaveButtonLabel": "确认保存",
     "common.ui.savedObjects.saveModal.duplicateTitleDescription": "单击 “{confirmSaveLabel}” 以保存标题重复的{objectType}",
